+++ src/resources/views/promotion-action/_form.blade.php
@@ -3,6 +3,7 @@
     <div class="col-md-10">
         {{ Form::select('type', $types, null, [
                 'class' => 'form-select form-select-sm' . ($errors->has('type') ? ' is-invalid': ''),
+                'id' => 'promotionActionTypeSelector',
                 'placeholder' => __('--')
            ])
         }}
@@ -14,17 +15,6 @@
 
 <hr>
 
-<div class="mb-3 row">
-    <label class="col-form-label col-form-label-sm col-md-2">{{ __('Configuration') }}</label>
-    <div class="col-md-10">
-        <textarea name="configuration"
-                  class="form-control form-control-sm{{ $errors->has('configuration') ? ' is-invalid' : '' }}"
-                  placeholder="{{ __('Enter JSON config') }}"
-                  rows="6"
-        >{{ old('configuration') ?? json_encode(Form::getModel()->configuration ?? [], JSON_PRETTY_PRINT | JSON_FORCE_OBJECT) }}</textarea>
-        @if ($errors->has('configuration'))
-            <div class="invalid-feedback">{{ $errors->first('configuration') }}</div>
-        @endif
-    </div>
-</div>
+<x-vanilo::configuration :model="$action" reload-on-change-of="promotionActionTypeSelector"
+                         :sample-refresh-route="route('vanilo.admin.promotion.action.create', $promotion)"></x-vanilo::configuration>