+++ src/resources/views/shipping-method/_form.blade.php
@@ -73,6 +73,7 @@
     <div class="col-md-10">
         {{ Form::select('calculator', $calculators, null, [
                 'class' => 'form-select form-select-sm' . ($errors->has('calculator') ? ' is-invalid': ''),
+                'id' => 'calculatorSelector',
                 'placeholder' => __('--')
            ])
         }}
@@ -82,17 +83,5 @@
     </div>
 </div>
 
-<div class="mb-3 row">
-    <label class="col-form-label col-form-label-sm col-md-2">{{ __('Configuration') }}</label>
-    <div class="col-md-10">
-        <textarea name="configuration"
-            class="form-control form-control-sm{{ $errors->has('configuration') ? ' is-invalid' : '' }}"
-            placeholder="{{ __('Enter JSON config') }}"
-            rows="6"
-        >{{ old('configuration') ?? json_encode(Form::getModel()->configuration ?? [], JSON_PRETTY_PRINT | JSON_FORCE_OBJECT) }}</textarea>
-        @if ($errors->has('configuration'))
-            <div class="invalid-feedback">{{ $errors->first('configuration') }}</div>
-        @endif
-    </div>
-</div>
-
+<x-vanilo::configuration :model="$shippingMethod" reload-on-change-of="calculatorSelector" pass-on-reload="calculator"
+     :sample-refresh-route="route('vanilo.admin.shipping-method.create')"></x-vanilo::configuration>