+++ Changelog.md
@@ -7,6 +7,7 @@
 - Added the Promotions (incl. Rules/Actions) & Coupons CRUD
 - Moved the Orders menu item to the top in the Shop section
 - Added input validation at Property Value assignment
+- Added the configuration widget, that can handle JSON config samples
 
 ## 4.1.1
 ##### 2024-07-23